Adding the given files for part 3 of the assignment and started on the class Buzam and the implementation of some Expressions

+++ src-provided/jumpingalien/part2/facade/Facade.java
@@ -377,7 +377,7 @@ public class Facade implements IFacadePart2 {
 	 *         should not belong to a world.
 	 */
 	public Plant createPlant(int x, int y, Sprite[] sprites) {
-		return new Plant((double) x,(double) y,sprites);
+		return new Plant(x,y,sprites);
 	}
 
 	/**
